Make code PSR-4 compliant
