+++ composer.json
@@ -9,15 +9,20 @@
 			"homepage": "http://wamania.com"
 		}
 	],
-    "require": {
+	"require": {
 		"php": ">=5.3.0"
-    },
-    "require-dev":{
-        "phpunit/phpunit": "^4.8"
-    },
+	},
+	"require-dev":{
+		"phpunit/phpunit": "^4.8"
+	},
 	"autoload": {
-        "psr-0": {
-            "Wamania": "src/"
-        }
-    }
+		"psr-4": {
+			"Wamania\\Snowball\\": "src/"
+		}
+	},
+	"autoload-dev": {
+		"psr-4": {
+			"Wamania\\Snowball\\Tests\\": "test/"
+		}
+	}
 }
