+++ src/Norwegian.php
