+++ src/Spanish.php
