+++ src/Stem.php
