+++ src/Stemmer.php
