+++ src/Swedish.php
